[Chore #161431131] Implement methods on Product model
